Refactoring.
Changed names of classes and files to match the convention.
Added slots and signals for Graphics items.
Moved doubleclick handling to the DrawingArea.
Polygon::addBoundaryPoint gets now QPointF instead of whole Polydot
object. So that it Polydot is now internal to the Polygon.
Polygon doesn't contain a list of it's inner polies. Now area holds this
information. It's better this way.